requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
cairocffi==1.7.1
cairosvg==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cssselect2==0.7.0
debugpy==1.8.12
decorator==5.1.1
deepsearch-glm==1.0.0
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
docling==2.22.0
docling-core==2.18.1
docling-ibm-models==3.3.2
docling-parse==3.3.1
easyocr==1.7.2
et-xmlfile==2.0.0
executing==2.2.0
filelock==3.17.0
filetype==1.2.0
fsspec==2025.2.0
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jinja2==3.1.5
jiter==0.8.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
kaleido==0.2.1
lancedb==0.19.0
latex2mathml==3.77.0
lazy-loader==0.4
lxml==5.3.1
markdown-it-py==3.0.0
marko==2.1.2
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multiprocess==0.70.17
narwhals==1.26.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.3
openai==1.63.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.1.0
pip==25.0
platformdirs==4.3.6
pnglatex==1.1
prompt-toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
pydeck==0.9.1
pygments==2.19.1
pylance==0.23.0
pynvim==0.5.2
pypdfium2==4.30.1
pyperclip==1.9.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rtree==1.3.0
safetensors==0.5.2
scikit-image==0.25.1
scipy==1.15.1
semchunk==2.2.2
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.42.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2025.1.10
tiktoken==0.9.0
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
typer==0.12.5
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
xlsxwriter==3.2.2
